% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{add_to_tibble}
\alias{add_to_tibble}
\title{Add Rows to a Tibble}
\usage{
add_to_tibble(tbl, ...)
}
\arguments{
\item{tbl}{A tibble to which new rows will be added.}

\item{...}{Variable arguments used to create a new tibble.}
}
\value{
The original tibble with the new rows appended at the end.
}
\description{
This function creates a new tibble from provided arguments and appends it to an existing tibble.
}
\examples{
# Assume `tbl` is an existing tibble
add_to_tibble(tbl, col1 = "value1", col2 = "value2")
}
\author{
Vincent Toups
}
