% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{add_where_clause_context}
\alias{add_where_clause_context}
\title{Add Context to Check Messages}
\usage{
add_where_clause_context(
  df,
  variable,
  where_clause_variable,
  where_clause_variable_value
)
}
\arguments{
\item{df}{A data frame containing at least the columns check_name and message.}

\item{variable}{The name of the variable adding context.}

\item{where_clause_variable}{The name of the variable that is part of the where_clause condition.}

\item{where_clause_variable_value}{The value of the where_clause variable.}
}
\value{
The same data frame, but with additional context added to the check_name and message columns.
}
\description{
This function adds additional context to the check_name and message columns of a data frame.
The context is defined by a given variable and its associated where_clause variable and value.
}
\examples{
df <- data.frame(check_name = "Check 1", message = "Error in row 1")
add_whereclause_context(df, "var1", "cond1", "value1")
}
\author{
Vincent Toups
}
