% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{apply_replace_quotes}
\alias{apply_replace_quotes}
\title{Apply replace_unicode_quotes to All Character Columns of a Tibble}
\usage{
apply_replace_quotes(tbl)
}
\arguments{
\item{tbl}{a tibble}
}
\value{
a tibble where all Unicode single and double quotation marks in all
character columns have been replaced by ASCII equivalents.
}
\description{
This function applies the function replace_unicode_quotes to each character
column of a tibble.
}
\examples{
data <- tibble::tibble(
  x = c("‘Hello, world!’", "“Hello, world!”"),
  y = c("‘Goodbye, world!’", "“Goodbye, world!”")
)
apply_replace_quotes(data)
# A tibble: 2 x 2
#   x                y               
#   <chr>            <chr>           
# 1 'Hello, world!'  'Goodbye, world!'
# 2 "Hello, world!"  "Goodbye, world!"
}
