% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.r
\name{args_ofat}
\alias{args_ofat}
\title{One Factor At a Time Argument Set}
\usage{
args_ofat(...)
}
\arguments{
\item{...}{Named arguments to a function.}
}
\value{
A tibble of argument combinations.
}
\description{
Generate an argument table based on OFAT.
}
\examples{
args_ofat(x = 1:5, y = 1:3)
args_ofat(x = 1:3, y = 1:3, z = 1:3)

}
