% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.r
\name{args_permute}
\alias{args_permute}
\title{Permutation Argument Set}
\usage{
args_permute(..., .n)
}
\arguments{
\item{...}{Named arguments to a function.}

\item{.n}{the number of argument permutations to evaluate
(sampling without replacement). If missing, all possible
permutations are returned.}
}
\value{
A tibble of argument combinations.
}
\description{
Generate an argument table based on permutations.
}
\examples{
args_permute(x = 1:5, y = 1:2)
args_permute(x = 1:10, y = 1:10, z = 1:10, .n = 10)

}
