% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.r
\name{args_set}
\alias{args_set}
\title{Argument Set}
\usage{
args_set(...)
}
\arguments{
\item{...}{Named arguments to a function.}
}
\value{
A tibble of argument combinations.
}
\description{
Generate an argument table from a set of arguments, following
the standard rules for vector recycling in R.
}
\examples{
args_set(x = 1:10, y = 1:10)
args_set(x = 1:10, y = 1:5, z = 1:2)
# mismatched argument lengths will generate a warning
\dontrun{
args_set(x = 1:10, y = 1:3)
}

}
