% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_codelist}
\alias{check_codelist}
\title{Check Codelist}
\usage{
check_codelist(
  dataset,
  variable_name,
  codelist,
  allow_na = TRUE,
  codelist_name = stop("You must specify a codelist name.")
)
}
\arguments{
\item{dataset}{A tibble/data.frame containing the data to be checked.}

\item{variable_name}{The name of the column to check.}

\item{codelist}{A vector of valid values for the column.}

\item{allow_na}{Logical indicating whether NA values should be allowed in the column. Default is TRUE.}
}
\value{
A tibble containing information about the checks and potential mismatches.
}
\description{
This function verifies if all the values of a specified column of a dataset are found in a given codelist.
It generates messages detailing checks and any mismatches. If a value is not in the codelist, it is considered an error.
}
\examples{
check_codelist(df, "column_name", c("value1", "value2"), allow_na = TRUE)
}
\author{
Vincent Toups
}
