% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_ftstat_values}
\alias{check_ftstat_values}
\title{Check that FTSTAT, where it does exist, is "Not done".}
\usage{
check_ftstat_values(ft_data)
}
\arguments{
\item{ft_data}{The data frame containing the ft dataset.}
}
\value{
A tibble summarizing the results of the check.
}
\description{
Check that FTSTAT, where it does exist, is "Not done".
}
