% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_iso8601_date}
\alias{check_iso8601_date}
\title{Check if values in a column can be parsed as ISO8601 dates}
\usage{
check_iso8601_date(dataset, variable_name, format = NULL, allow_na = FALSE)
}
\arguments{
\item{dataset}{A data frame containing the data to be checked.}

\item{variable_name}{The name of the variable/column to check.}

\item{format}{A specification of the format that the date values should conform to. Must be "ISO8601".}

\item{allow_na}{Logical flag indicating whether NA values are allowed. Defaults to FALSE.}
}
\value{
A tibble that records the name of the check, whether it passed, a message summarizing the check results, and the row numbers of any values that failed the check.
}
\description{
Check if values in a column can be parsed as ISO8601 dates
}
\examples{
\dontrun{
check_iso8601_date(dataset = df, variable_name = "column1", format = "ISO8601", allow_na = FALSE)
}
}
