% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_ISO8601_durations_cpp}
\alias{check_ISO8601_durations_cpp}
\title{Check if a vector of strings are valid ISO8601 durations}
\usage{
check_ISO8601_durations_cpp(putative_duration_strings)
}
\arguments{
\item{putative_duration_strings}{A vector of strings that may represent ISO8601 durations.}
}
\value{
A LogicalVector of the same length as the input vector.
Each element in the returned vector indicates whether the corresponding string in
the input vector is a valid ISO8601 duration (TRUE) or not (FALSE).
NAs in the input vector are preserved in the returned vector.
}
\description{
This function takes a vector of strings as input, and returns a LogicalVector
indicating whether each string is a valid ISO8601 duration.
For each string in the input vector, the function checks if it is a valid ISO8601 duration.
If the string is a valid duration, it returns TRUE; otherwise, it returns FALSE.
If the string is NA, the function returns NA.
}
\examples{
# Assume `reval` is the name of your package
library(reval)

durations = c("P1Y2M", "INVALID", "P10D", NA)
check_ISO8601_durations_cpp(durations)
}
