% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_missing_values_in_ft}
\alias{check_missing_values_in_ft}
\title{Check that FTSTRESC, FTSTRESN, and FTSTAT are not all missing.}
\usage{
check_missing_values_in_ft(ft_data)
}
\arguments{
\item{ft_data}{The data frame containing the ft dataset.}
}
\value{
A tibble summarizing the results of the check.
}
\description{
Check that FTSTRESC, FTSTRESN, and FTSTAT are not all missing.
}
