% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_missing_values_in_pairs}
\alias{check_missing_values_in_pairs}
\title{Check for rows where both columns in each name pair are missing and create a boolean array.}
\usage{
check_missing_values_in_pairs(data)
}
\arguments{
\item{data}{The data frame to check.}
}
\value{
A list containing a tibble summarizing the results of the check and a boolean array.
}
\description{
This function identifies pairs of columns that differ only by their last character.
It then checks if both columns in the pair are missing for the same row.
A summary of the check is returned in a tibble.
Pairs of columns where either name starts with 'FT' are ignored.
A boolean array where every index in problematic_rows is 0 and all other indices are 1 is computed.
}
