% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_na}
\alias{check_na}
\title{Check for NA Values in a Column of a Dataset}
\usage{
check_na(dataset, variable_name)
}
\arguments{
\item{dataset}{A tibble or dataframe containing the data to be checked.}

\item{variable_name}{A string that specifies the column name to be checked for NA values.}
}
\value{
A tibble that includes the check name (specified as "Column \link{variable_name} doesn't have NA values."),
a boolean indicating whether the check passed, an associated message, and the row numbers where NA values were found.
The row numbers are formatted using the \code{format_row_numbers} function.
}
\description{
This function checks if a column in a dataset has any NA (Not Available) values.
It generates a tibble that includes the check name, whether the check passed,
the associated message, and the row numbers where NA values were found.
}
\examples{
# Assuming 'df' is a dataset and 'age' is one of its columns
check_na(df, "age")

 
}
