% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_numeric_codelist}
\alias{check_numeric_codelist}
\title{Check Numeric Codelist}
\usage{
check_numeric_codelist(
  dataset,
  variable_name,
  codelist,
  allow_na = TRUE,
  codelist_name
)
}
\arguments{
\item{dataset}{A tibble/data.frame containing the data to be checked.}

\item{variable_name}{The name of the column to check.}

\item{codelist}{A numeric vector of valid values for the column.}

\item{allow_na}{Logical indicating whether NA values should be allowed in the column. Default is TRUE.}
}
\value{
A tibble containing information about the checks and potential mismatches.
}
\description{
This function is called by check_codelist when the codelist is numeric.
It converts the codelist and the column values to numeric before running
the check_codelist function again.
}
\examples{
check_numeric_codelist(df, "column_name", c("1", "2"), allow_na = TRUE)
}
\author{
Vincent Toups
}
