% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_qsevlint}
\alias{check_qsevlint}
\title{Check compliance of QSEVLINT column in a dataset with spec}
\usage{
check_qsevlint(dataset, combined_spec)
}
\arguments{
\item{dataset}{A data frame that includes columns 'QSEVLINT' and 'QSTESTCD'.}

\item{combined_spec}{A data frame that includes the specifications for the columns, including
'codelist_terms' and 'evaluation_interval'. Rows where 'evaluation_interval' is NA are ignored.}
}
\value{
A tibble containing the name of the check, whether it passed, a message indicating the result,
and the row numbers in the original dataset that failed the check (if any).
}
\description{
This function checks whether the 'QSEVLINT' values in a dataset are consistent with their corresponding
'QSTESTCD' values, as specified in a combined spec.
}
\examples{
spec <- tibble::tribble(
  ~codelist_terms, ~evaluation_interval,
  "QSEVLINT_1",         "1",
  "QSEVLINT_2",         "2",
  "QSEVLINT_3",         NA
)

data <- tibble::tribble(
  ~QSEVLINT, ~QSTESTCD,
  "QSEVLINT_1", "1",
  "QSEVLINT_2", "2",
  "QSEVLINT_3", "3"
)

check_qsevlint(data, spec)

}
