% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_text}
\alias{check_text}
\title{Check if values in a column are textual, with optional NA check}
\usage{
check_text(dataset, variable_name, spec_type, allow_na = FALSE)
}
\arguments{
\item{dataset}{A data frame containing the data to be checked.}

\item{variable_name}{The name of the variable/column to check.}

\item{allow_na}{Logical flag indicating whether NA values are allowed. Defaults to FALSE.}
}
\value{
A tibble that records the name of the check, whether it passed, a message summarizing the check results, and the row numbers of any values that failed the check.
}
\description{
Check if values in a column are textual, with optional NA check
}
\examples{
\dontrun{
check_text(dataset = df, variable_name = "column1", allow_na = FALSE)
}
}
