% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_type}
\alias{check_type}
\title{Check the type of values in a column based on provided specification}
\usage{
check_type(dataset, variable_name, spec_type, format = NULL, allow_na = FALSE)
}
\arguments{
\item{dataset}{A data frame containing the data to be checked.}

\item{variable_name}{The name of the variable/column to check.}

\item{spec_type}{The expected type of the values in the column. Must be one of "integer", "float", "text", "date", "duration".}

\item{format}{An optional specification of the format that the values should conform to.}

\item{allow_na}{Logical flag indicating whether NA values are allowed. Defaults to FALSE.}
}
\value{
A tibble that records the name of the check, whether it passed, a message summarizing the check results, and the row numbers of any values that failed the check.
}
\description{
Check the type of values in a column based on provided specification
}
\examples{
\dontrun{
check_type(dataset = df, variable_name = "column1", spec_type = "integer", format = NULL, allow_na = FALSE)
}
}
