% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{check_where_clause}
\alias{check_where_clause}
\title{Apply WHERE Clause Checks}
\usage{
check_where_clause(dataset, spec, dataset_name, variable_name)
}
\arguments{
\item{dataset}{A data frame representing the dataset to be checked.}

\item{spec}{A data frame containing the specification of the dataset variables.}

\item{dataset_name}{A character string specifying the name of the dataset.}

\item{variable_name}{A character string specifying the name of the variable.}
}
\value{
A data frame containing information about the variable for each subset of data defined by the WHERE clause.
}
\description{
This function applies a set of checks, based on a given specification, to a data column within a dataset.
The data column is assumed to have a WHERE clause, meaning it needs to be evaluated differently depending on the values of another column.
This function groups the data by that other column's values and applies the appropriate checks for each group.
}
\examples{
# Assuming a dataset `data` and its specification `spec`
check_where_clause(data, spec, "dataset_name", "variable_name")
}
\author{
Vincent Toups
}
