% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{collapse_commas}
\alias{collapse_commas}
\title{Collapses a Character Vector into a Comma-Separated String}
\usage{
collapse_commas(s)
}
\arguments{
\item{s}{a character vector}
}
\value{
a single string with elements of the input vector separated by commas
}
\description{
This function takes a character vector as input and collapses it into a single
string, with elements separated by commas.
}
\examples{
collapse_commas(c("apple", "banana", "cherry"))
# [1] "apple, banana, cherry"

}
