% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{collapse_newlines_tab}
\alias{collapse_newlines_tab}
\title{Collapses a Character Vector into a Newline-Tab-Separated String}
\usage{
collapse_newlines_tab(s)
}
\arguments{
\item{s}{a character vector}
}
\value{
a single string with elements of the input vector separated by newlines followed by a tab.
}
\description{
This function takes a character vector as input and collapses it into a single
string, with elements separated by newlines followed by a tab.
}
\examples{
collapse_newlines_tab(c("apple", "banana", "cherry"))
# [1] "apple\n\tbanana\n\tcherry"

}
