% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reval.r
\name{collate_many}
\alias{collate_many}
\title{Collate results of many evaluations}
\usage{
collate_many(l, reshape.fun = identity, ..., variable.name = "variable",
  value.name = "value")
}
\arguments{
\item{l}{The list of lists of vectors or dataframes,
e.g. output of \code{eval_many}. Tries to coerce each element into a
dataframe with \code{as.data.frame} before collating.}

\item{reshape.fun}{A function for reshaping each element in \code{l} before
collating.}

\item{...}{Other arguments to be passed to \code{reshape.fun}.}

\item{variable.name}{The name of the column used to store the variable names.}

\item{value.name}{The name of the column used to store the variable values.}
}
\value{
A dataframe collated from \code{l}.
}
\description{
Collate a list of lists of vectors or dataframes into one dataframe.
}
\details{
Loops through each element in each list, ignoring \code{NULL}
  elements. For complex elements, the argument \code{reshape.fun} can
  be used to extract information (see examples).
}
\examples{
res = eval_many(rnorm, mean=c(5,10), sd=c(2, 3), default.args=list(n=5))

collate_many(res)
collate_many(res, reshape.fun=t, variable.name="parameter",
  value.name = "parameter.value")

\dontrun{
datasets = list(
  A = data.frame(x=1:10, y = 1:10 + rlnorm(10)),
  B = data.frame(x=1:10, y = 1:10 + rnorm(10, mean=2))
)
res = eval_many(lm, data=datasets, default.args=list(formula = y ~ x))
collate_many(res, reshape.fun = function(x) t(x$coefficients))
}
}

