% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{convert_yes_no_to_logical}
\alias{convert_yes_no_to_logical}
\title{Convert "Yes" and "No" to Logical Values}
\usage{
convert_yes_no_to_logical(char_vec)
}
\arguments{
\item{char_vec}{A character vector that should contain only "Yes" and "No" values to convert to logical.}
}
\value{
A logical vector with \code{TRUE} for "Yes" and \code{FALSE} for "No".
}
\description{
This function takes a character vector as input and converts elements with the value "Yes" to \code{TRUE}
and elements with the value "No" to \code{FALSE}. If any other values are encountered in the vector,
the function will stop execution and return an informative error message.
}
