% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{debug_print}
\alias{debug_print}
\title{Print debugging information conditionally}
\usage{
debug_print(x, ...)
}
\arguments{
\item{x}{The object to be printed.}

\item{...}{Further arguments passed to the print function.}
}
\description{
This function behaves like the base print function, but only prints the output
when the package-level variable 'debugging' is TRUE.
}
