% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\docType{data}
\name{debugging}
\alias{debugging}
\title{Debugging status variable}
\format{
An object of class \code{logical} of length 1.
}
\usage{
debugging
}
\description{
Debugging status variable
}
\keyword{internal}
