% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\docType{data}
\name{error_context}
\alias{error_context}
\title{Package-level variable for error message context}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
error_context
}
\description{
This is a package level variable that provides extra context
for error messages generated by this library. It is intended
to be temporarily set by \code{with_error_context()} function to
augment error messages.
}
\keyword{datasets}
