% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reval.r
\name{eval_many}
\alias{eval_many}
\title{Repeated evaluations along one dimension}
\usage{
eval_many(fun, ..., method = "OFAC", default.args = list())
}
\arguments{
\item{fun}{The function to be evaluated.}

\item{...}{Arguments to be varied when evaluating \code{fun}, where each
argument in \code{...} is a vector or list of values.}

\item{method}{The sensitivity analysis method to be used. Currently supports
one-factor-at-a-time ("OFAC") evaluation.}

\item{default.args}{the default values of arguments passed to \code{fun}.}
}
\value{
A list of lists: The top-level names are the arguments in \code{...}
  and the bottom-level names are the names of each element in each argument
  in \code{...}.
}
\description{
Evaluate a function repeatedly, varying one argument at a time.
}
\details{
If the elements of an argument in \code{...} are not named, the
  values of the elements will used via \code{paste}.
}
\examples{
eval_many(rnorm, mean=c(5,10), sd=c(2, 3), default.args=list(n=5))
eval_many(rnorm, mean=c(5,10), sd=c(2, 3), default.args=list(n=5, mean=20, sd=5))

\dontrun{
formulas = list(y ~ x, y ~ x^2, y ~ x + z)
datasets = list(
  A = data.frame(x=1:10, y = 1:10 + rlnorm(10)),
  B = data.frame(x=1:10, y = 1:10 + rnorm(10, mean=2))
)
eval_many(lm, data=datasets, default.args=list(formula = formulas[[1]]))
eval_many(lm, formula=formulas, default.args=list(datasets$B))
}
}

