% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{filter_rows_all_na}
\alias{filter_rows_all_na}
\title{Filter rows where all specified columns have NA values}
\usage{
filter_rows_all_na(df, cols)
}
\arguments{
\item{df}{A data frame containing the data to be filtered.}

\item{cols}{A character vector of column names that should be checked for NA values.}
}
\value{
A data frame with rows removed where all specified columns are NA.
}
\description{
This function filters out the rows in a data frame where all values in the specified columns are NA.
}
\examples{
# Create an example data frame
df <- tibble(
  a = c(1, 2, NA, NA, 5),
  b = c(NA, 2, 3, NA, 5),
  c = c(1, NA, NA, NA, 5)
)

# Filter out rows where all of columns 'a', 'b', and 'c' are NA
filtered_df <- filter_rows_all_na(df, c("a", "b", "c"))

}
