% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{find_name_pairs}
\alias{find_name_pairs}
\title{Find pairs of variables with names differing only by the last character being "C" or "N"}
\usage{
find_name_pairs(names)
}
\arguments{
\item{names}{A character vector of variable/column names.}
}
\value{
A list of character vectors, each of length 2, representing pairs of variable names where the names are identical except for the last character, which is "C" in one name and "N" in the other.
}
\description{
Find pairs of variables with names differing only by the last character being "C" or "N"
}
\examples{
\dontrun{
find_name_pairs(names = c("variable1C", "variable1N", "variable2", "variable3C"))
}
}
