% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{fix_format}
\alias{fix_format}
\title{Fix Numeric Format Specifications}
\usage{
fix_format(format)
}
\arguments{
\item{format}{A vector of character strings representing the formats to be fixed.}
}
\value{
A vector of character strings with the fixed formats.
}
\description{
This function fixes numeric format specifications. Specifically, it is designed to deal with issues
that arise due to floating-point precision when serializing and deserializing numbers. The function takes
a vector of format specifications as input. If a specification includes a decimal, the decimal part is
rounded to the nearest integer. Specifications without a decimal are returned as is.
}
