% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{fix_format_helper}
\alias{fix_format_helper}
\title{Internal Helper Function for Fixing Formats}
\usage{
fix_format_helper(format)
}
\arguments{
\item{format}{A character string representing the format to be fixed.}
}
\value{
A character string with the fixed format.
}
\description{
This function is used internally by fix_format() function. It is not intended to be used directly by users.
}
\keyword{internal}
