% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{format_row_numbers}
\alias{format_row_numbers}
\title{Format Row Numbers}
\usage{
format_row_numbers(row_numbers, max_n = 10)
}
\arguments{
\item{row_numbers}{A vector of row numbers to be formatted.}

\item{max_n}{Maximum number of row numbers to be included in the formatted string.}
}
\value{
A formatted string of row numbers.
}
\description{
This function formats a vector of row numbers into a comma-separated string.
If the length of the vector is larger than \code{max_n}, it will be truncated and " ..." will be appended at the end.
}
\examples{
format_row_numbers(c(1,2,3,4,5), max_n = 3)
}
\author{
Vincent Toups
}
