% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{get_codelist_name}
\alias{get_codelist_name}
\title{Get Codelist Name}
\usage{
get_codelist_name(variable_name, spec)
}
\arguments{
\item{variable_name}{A character string specifying the name of the variable.}

\item{spec}{A data frame containing the specification of the dataset variables.}
}
\value{
A character vector containing unique codelist name.
}
\description{
Get the unique codelist name from the specification for a given variable.
}
\examples{
# Assuming a dataset specification `spec`
get_codelist_name("variable", spec)
}
\author{
Vincent Toups
}
