% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{get_codelist_terms}
\alias{get_codelist_terms}
\title{Get Codelist Terms}
\usage{
get_codelist_terms(variable_name, spec)
}
\arguments{
\item{variable_name}{A character string specifying the name of the variable.}

\item{spec}{A data frame containing the specification of the dataset variables.}
}
\value{
A character vector containing unique codelist terms.
}
\description{
Get the unique codelist terms from the specification for a given variable.
}
\examples{
# Assuming a dataset specification `spec`
get_codelist_terms("variable", spec)
}
\author{
Vincent Toups
}
