% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{get_format}
\alias{get_format}
\title{Get Format Field}
\usage{
get_format(dataset_name, variable_name, spec)
}
\arguments{
\item{dataset_name}{A character string specifying the name of the dataset.}

\item{variable_name}{A character string specifying the name of the variable.}

\item{spec}{A data frame containing the specification of the dataset variables.}
}
\value{
A character string indicating the format of the variable.
}
\description{
Get the format for a variable in a dataset according to a provided specification.
}
\examples{
# Assuming a dataset specification `spec` and a dataset `dataset`
get_format("dataset", "variable", spec)
}
\author{
Vincent Toups
}
