% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{get_variable_type}
\alias{get_variable_type}
\title{Get Variable Type}
\usage{
get_variable_type(dataset_name, variable_name, spec)
}
\arguments{
\item{dataset_name}{A character string specifying the name of the dataset.}

\item{variable_name}{A character string specifying the name of the variable.}

\item{spec}{A data frame containing the specification of the dataset variables.}
}
\value{
A character string indicating the type of the variable.
}
\description{
Get the type of a variable in a dataset according to a provided specification.
}
\examples{
# Assuming a dataset specification `spec` and a dataset `dataset`
get_variable_type("dataset", "variable", spec)
}
\author{
Vincent Toups
}
