% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{is_numeric_codelist}
\alias{is_numeric_codelist}
\title{Check if a character vector can be parsed entirely as numeric}
\usage{
is_numeric_codelist(x)
}
\arguments{
\item{x}{A character vector to be checked.}
}
\value{
A logical value indicating whether all elements in the vector can be parsed as numeric.
}
\description{
This function checks whether all elements in a character vector can be successfully
parsed as numeric values. If all elements can be parsed as numeric, the function
returns TRUE; otherwise, it returns FALSE.
}
\examples{
is_numeric_codelist(c("1", "2", "3"))  # TRUE
is_numeric_codelist(c("1", "2", "a"))  # FALSE

}
