% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{load_and_combine_specs}
\alias{load_and_combine_specs}
\title{Load and combine a set of dataset specifications.}
\usage{
load_and_combine_specs(
  spec_directory,
  sheet_names = c("Study", "Datasets", "Variables", "ValueLevel", "WhereClauses",
    "Codelists", "Dictionaries", "Methods", "Comments", "Documents")
)
}
\arguments{
\item{spec_directory}{A string specifying the directory in which the Excel and ODS files are located.}

\item{sheet_names}{A character vector specifying the names of the sheets to be read from the files.}
}
\value{
A named list of data frames. The names correspond to the sheet names, and the data frames contain the data read from the corresponding sheets of the files.
}
\description{
Load and combine a set of dataset specifications.
}
\examples{
\dontrun{
create_list_of_df(spec_directory = "path/to/directory", sheet_names = c("Sheet1", "Sheet2"))
}
}
