% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{merge_spec_files}
\alias{merge_spec_files}
\title{Merge specification files}
\usage{
merge_spec_files(
  spec_directory,
  sheet_names = c("Study", "Datasets", "Variables", "ValueLevel", "WhereClauses",
    "Codelists", "Dictionaries", "Methods", "Comments", "Documents")
)
}
\arguments{
\item{spec_directory}{String. The directory where the specification files are located.}

\item{sheet_names}{Character vector. The names of the sheets to be merged.}
}
\value{
A combined and digested data frame with all the information from the specs required to check the covered data domains.
}
\description{
This function merges multiple specification files from a directory
into a single data frame for each sheet and the combines the result
into a format convenient for validation.
}
