% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval-package.R
\docType{data}
\name{min_data_set_spec}
\alias{min_data_set_spec}
\title{The latest digested combined minimum data set specification at
the time of this package's last build.}
\format{
The exact format of this data is an implementation
detail. See the function reval::merge_spec_files, which allows
the user to take a directory of downloaded MDS specifications
and produce the appropriate data frame if they want to use a
newer or otherwise modified form of the specification.
}
\source{
Contact the BACPAC DAC.
}
\usage{
min_data_set_spec
}
\description{
This data frame contains sufficient information to determine the
the checks that need to be performed on each variable in a data
set. It is a combination of the pages found in the specifications
distributed by BACPAC's DAC.
}
\keyword{datasets}
