% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{parses_as_float}
\alias{parses_as_float}
\title{Parse as Float}
\usage{
parses_as_float(values)
}
\arguments{
\item{values}{A character vector to be checked.}
}
\value{
A logical vector indicating whether each value can be parsed as a float.
}
\description{
This function checks if given values can be parsed as floating point numbers.
}
