% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{parses_as_formatted_float}
\alias{parses_as_formatted_float}
\title{Parse as Formatted Float}
\usage{
parses_as_formatted_float(input_string, format)
}
\arguments{
\item{input_string}{A string to be checked.}

\item{format}{A string specifying the maximum number of digits before and after the decimal point.}
}
\value{
A logical value indicating whether the string can be parsed as a formatted float.
}
\description{
This function checks if a given string can be parsed as a float with a specified format.
}
