% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{parses_as_formatted_int}
\alias{parses_as_formatted_int}
\title{Parse as Formatted Integer}
\usage{
parses_as_formatted_int(input_string, format)
}
\arguments{
\item{input_string}{A string to be checked.}

\item{format}{A string specifying the maximum number of digits the integer can have.}
}
\value{
A logical value indicating whether the string can be parsed as a formatted integer.
}
\description{
This function checks if a given string can be parsed as an integer with a specified maximum number of digits.
}
