% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{parses_as_int}
\alias{parses_as_int}
\title{Parse as Integer}
\usage{
parses_as_int(values)
}
\arguments{
\item{values}{A character vector to be checked.}
}
\value{
A logical vector indicating whether each value can be parsed as an integer.
}
\description{
This function checks if given values can be parsed as integers.
}
