% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{parses_as_iso8601_date}
\alias{parses_as_iso8601_date}
\title{Check if a character vector contains valid ISO 8601 date strings}
\usage{
parses_as_iso8601_date(date_vec)
}
\arguments{
\item{date_vec}{Character vector with potential date strings.}
}
\value{
Logical vector indicating whether each string is a valid ISO 8601 date.
}
\description{
Check if a character vector contains valid ISO 8601 date strings
}
