% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{parses_as_iso8601_duration}
\alias{parses_as_iso8601_duration}
\title{Check if values can be parsed as ISO8601 durations}
\usage{
parses_as_iso8601_duration(duration_vec)
}
\arguments{
\item{duration_vec}{A character vector containing the duration strings to check.}
}
\value{
A logical vector with the same length as \code{duration_vec},
indicating whether each value can be parsed as an ISO8601 duration.
}
\description{
This function is a wrapper for the \code{check_iso8601_durations_cpp} function
implemented in C++. It checks if the input values can be parsed as durations
according to the ISO8601 standard.
}
\note{
The actual implementation of the parsing check is done by the
\code{check_iso8601_durations_cpp} function, which provides a more efficient
implementation using C++.
}
\examples{
\dontrun{
parses_as_iso8601_duration(c("P1Y2M3DT4H5M6S", "invalid"))
}

}
\seealso{
\code{\link{check_iso8601_durations_cpp}}
}
