% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval-package.R
\docType{data}
\name{qsmd_test_data}
\alias{qsmd_test_data}
\title{An artificially created QSMD test dataset.}
\format{
see QSMDS Specification
}
\source{
Contact the BACPAC DAC.
}
\usage{
qsmd_test_data
}
\description{
An artificially created QSMD test dataset.
}
\keyword{datasets}
