% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{future_pmap}
\alias{plan}
\alias{sequential}
\alias{transparent}
\alias{multisession}
\alias{multicore}
\alias{cluster}
\alias{remote}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{furrr}{\code{\link[furrr:future_map2]{future_pmap}}}

  \item{future}{\code{\link[future]{cluster}}, \code{\link[future]{multicore}}, \code{\link[future]{multisession}}, \code{\link[future]{plan}}, \code{\link[future]{remote}}, \code{\link[future]{sequential}}, \code{\link[future:sequential]{transparent}}}
}}

