% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{replace_unicode_quotes}
\alias{replace_unicode_quotes}
\title{Replace Unicode Quotation Marks in a String}
\usage{
replace_unicode_quotes(input_str)
}
\arguments{
\item{input_str}{a character string}
}
\value{
a character string where all Unicode single and double quotation marks
have been replaced by ASCII equivalents.
}
\description{
This function replaces all Unicode single and double quotation mark-like
characters with their ASCII equivalents in a character string.
}
\examples{
replace_unicode_quotes(c("‘Hello, world!’", "“Hello, world!”"))
# [1] "'Hello, world!'" '"Hello, world!"'
}
