% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{reval_read_csv}
\alias{reval_read_csv}
\title{Load a data frame but keep both a parsed and unparsed version of each column.}
\usage{
reval_read_csv(filename, col_types = NULL)
}
\arguments{
\item{filename}{Path to the CSV file to be loaded.}

\item{col_types}{Column types. NULL by default.}
}
\value{
A tibble with twice the columns of the original file, half of which are unparsed duplicates.
}
\description{
Load a data frame but keep both a parsed and unparsed version of each column.
}
