% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{reval_read_data}
\alias{reval_read_data}
\title{Load either a CSV or XPT file and save unparsed rows}
\usage{
reval_read_data(filename)
}
\arguments{
\item{filename}{Path to the CSV or XPT file to be loaded.}
}
\value{
A tibble with both parsed and unparsed columns.
}
\description{
Load either a CSV or XPT file and save unparsed rows
}
