% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{reval_read_xpt}
\alias{reval_read_xpt}
\title{Load a data frame from an xpt but keep both a parsed and unparsed version of each column.}
\usage{
reval_read_xpt(filename)
}
\arguments{
\item{filename}{Path to the XPT file to be loaded.}
}
\value{
A tibble with twice the columns of the original file, half of which are unparsed duplicates.
}
\description{
Load a data frame from an xpt but keep both a parsed and unparsed version of each column.
}
