% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{unparsed_column_name}
\alias{unparsed_column_name}
\title{Return the unparsed column name for "name"}
\usage{
unparsed_column_name(name)
}
\arguments{
\item{name}{(or names)}
}
\value{
the unparsed name
}
\description{
Return the unparsed column name for "name"
}
