% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reval.R
\name{with_error_context}
\alias{with_error_context}
\title{Execute a block of code with additional error context}
\usage{
with_error_context(new_error_context, body, ...)
}
\arguments{
\item{new_error_context}{The new context message to set.}

\item{body}{The block of code to execute with the new error context.}

\item{...}{Additional arguments passed to the body function.}
}
\value{
The value returned by the body function.
}
\description{
This function sets a new error context, executes a given block of code,
and then restores the previous error context. This is useful for providing
more detailed error messages.
}
\examples{
\dontrun{
with_error_context("Reading data file", {
  # Your code here e.g., read.csv("some_file.csv")
})
}

}
