% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reval.r
\name{eval_many}
\alias{eval_many}
\title{Repeated evaluations}
\usage{
eval_many(fun, ..., method = c("ofac", "permute", "set"), sample = NA,
  default.args = list(), collate = TRUE, reshape.fun = identity,
  collate.id = c("single", "multi"))
}
\arguments{
\item{fun}{The function to be evaluated.}

\item{...}{Arguments to be varied when evaluating \code{fun}, where each
argument in \code{...} is a vector or list of values.}

\item{method}{The sensitivity analysis method to be used. Can be either
one-factor-at-a-time ("ofac") evaluation, evaluation of parameter sets
("set"), or (sampled) permutations of parameter sets ("permute"). When
\code{method = "ofac"}, the first element of each argument in \code{...}
is assumed to be the "default" value of that argument.}

\item{sample}{If \code{method = "permute"}, the number of parameter
permutations to evaluate. If NA, all permutations will be evaluated.}

\item{default.args}{the default values of arguments passed to \code{fun}.}

\item{collate}{Whether to collate the results or not. If TRUE, output
elements will be coerced into data.frames using \code{as.data.frame}.}

\item{reshape.fun}{A function for reshaping the output of each evaluation
prior to coercing and collating the outputs.}

\item{collate.id}{Only used when \code{collate = TRUE}. The method used to
  store the evaluation identifiers. If \code{collate.id = "single"}, a single
  column column is used. If \code{collate.id = "multi"}, one column is created
  for each argument in \code{...}. \code{collate.id = "multi"} can at times
  simplify data exploration and visualization.}
}
\value{
If \code{collate = TRUE}, a data.frame. Otherwise, a list structure
  depending on the value of \code{method}:
  \item{"ofac"}{A list of lists: The top-level names are
    the arguments in \code{...} and the bottom-level names are the names of
    each element of that argument. If the elements of an argument in
    \code{...} are not named, the values of the elements will used via
    \code{paste}.}
  \item{"set"}{A simple list.}
}
\description{
Evaluate a function repeatedly across argument sets or permutations.
}
\examples{
set.seed(1459)
eval_many(rnorm, mean = c(5, 9), sd = c(2, 3), default.args = list(n = 5))
eval_many(rnorm, mean = c(5, 9), sd = c(2, 3), default.args = list(n = 5),
  reshape.fun = t, collate.id = "multi")
eval_many(rnorm, mean=seq(20), sd = seq(1, 4, by = 0.1),
 default.args = list(n = 5), method = "permute", sample = 10,
  reshape.fun = t)

# vector recycling
eval_many(rnorm, mean = c(0, 5, 3), sd = c(0, 1), method = "set",
  default.args = list(n = 5), reshape.fun = t, collate.id = "multi")

\donttest{\dontrun{
formulas = list(y ~ x, y ~ x^2, y ~ x + z)
datasets = list(
  A = data.frame(x=1:10, y = 1:10 + rnorm(10)),
  B = data.frame(x=1:10, y = 1:10 + rnorm(10, mean = 2))
)

# raw output
eval_many(lm, formula = formulas, data = datasets[1], method = "set",
  collate = FALSE)

# data extraction
eval_many(lm, formula = formulas, data = datasets, method = "permute",
  reshape.fun = function(x) t(x$coefficients), collate.id = "multi")
}}
}

