% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reval.r
\docType{package}
\name{reval-package}
\alias{reval-package}
\title{This package provides the function \code{evalmany} to simplify scenario
testing and sensitivity analysis with R.}
\description{
This package provides the function \code{evalmany} to simplify scenario
testing and sensitivity analysis with R.
}

